CreatorChatCRUD/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
deepgram-sdk==2.0.0
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
ipdb==0.13.13
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
multidict==6.0.4
mutagen==1.47.0
openai==1.6.1
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
protobuf==4.25.1
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycryptodomex==3.19.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
stack-data==0.6.3
tomli==2.0.1
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.12
websockets==12.0
Werkzeug==3.0.1
yarl==1.9.4
yt-dlp==2023.11.16
zipp==3.17.0
